2.4.0 Update

+++ src/main/java/net/fneifnox/custommobattributes/mixin/damage/DragonPurgeMixin.java
@@ -9,13 +9,9 @@ import static net.fneifnox.custommobattributes.CustomMobAttributes.CONFIG;
 
 @Mixin(EnderDragonEntity.class)
 public class DragonPurgeMixin {
-    @ModifyArg(method = "damageLivingEntities", at = @At(
-            value = "INVOKE",
-            target = "Lnet/minecraft/entity/Entity;damage(Lnet/minecraft/entity/damage/DamageSource;F)Z"
-    )
-    )
+    @ModifyArg(method = "damageLivingEntities", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;damage(Lnet/minecraft/entity/damage/DamageSource;F)Z"))
     private float modifyDamageAmount(float originalDamage) {
-        float multiplier = CONFIG.damageMultiplierForEnderDragon() * CONFIG.damageMultiplierForAll();
-        return originalDamage * multiplier;
+        double multiplier = CONFIG.damageMultiplierForEnderDragon() * CONFIG.damageMultiplierForAll();
+        return (float) (originalDamage * multiplier);
     }
 }
