+++ src/main/java/net/fneifnox/custommobattributes/mixin/damage/SnowballMixin.java
@@ -12,21 +12,15 @@ import static net.fneifnox.custommobattributes.CustomMobAttributes.CONFIG;
 
 @Mixin(SnowballEntity.class)
 public class SnowballMixin {
-    @Redirect(
-            method = "onEntityHit",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/entity/Entity;damage(Lnet/minecraft/entity/damage/DamageSource;F)Z"
-            )
-    )
+    @Redirect(method = "onEntityHit", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;damage(Lnet/minecraft/entity/damage/DamageSource;F)Z"))
     private boolean redirectDamage(Entity entity, DamageSource source, float originalDamage) {
-        float multiplier = 1f;
+        double multiplier = 1f;
         SnowballEntity snowball = (SnowballEntity)(Object)this;
         if (snowball.getOwner() instanceof SnowGolemEntity) {
             multiplier = CONFIG.damageMultiplierForSnowGolem() * CONFIG.damageMultiplierForAll();
         }
 
-        float finalDamage = originalDamage * multiplier;
-        return entity.damage(source, finalDamage);
+        double finalDamage = originalDamage * multiplier;
+        return entity.damage(source, (float) finalDamage);
     }
 }
